+++ muk_ai_automation/tests/__init__.py
@@ -6,14 +6,9 @@ from . import test_action_agent_run
 from . import test_base_automation_path
 from . import test_caps
 from . import test_chain_strategy
-from . import test_chatter_mirror_access
 from . import test_client_kinds
 from . import test_cron_path
 from . import test_dispatch_errors
 from . import test_dispatch_helper
 from . import test_dispatch_modes
 from . import test_run_contract
-from . import test_session_record_acl
-from . import test_session_transcript_access
-from . import test_js
-from . import test_tour